Download Subtitles of Hollywood Movies
It downloads the subtitle of the movie. It first takes the path of the directory where you want to save the download .srt file. Make sure you enter it perfectly. Second, it takes name of the movie and downloads the subtitle of the movie in the background. If there are multiple subtitles of a movie then all options are listed and you have choose only one.